requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
blinker==1.9.0
build==1.2.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromadb==0.3.23
cleo==2.1.0
click==8.1.7
clickhouse-connect==0.7.8
colorama==0.4.6
crashtest==0.4.1
dataclasses-json==0.6.5
distlib==0.3.8
distro==1.9.0
Django==5.0.6
django-admin-list-filter-dropdown==1.0.3
django-bootstrap-email==0.0.1
django-bulma==0.8.3
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==24.2
django-modeladmin-reorder==0.3.1
django-widget-tweaks==1.5.0
duckdb==0.10.2
dulwich==0.21.7
fastapi==0.110.2
fastjsonschema==2.19.1
filelock==3.13.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
groq==0.5.0
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
hnswlib==0.8.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
keyring==24.3.1
langchain==0.1.16
langchain-cli==0.0.21
langchain-community==0.0.34
langchain-core==0.1.46
langchain-groq==0.1.3
langchain-nomic==0.0.2
langchain-text-splitters==0.0.1
langserve==0.1.1
langsmith==0.1.51
libclang==18.1.1
loguru==0.7.2
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.13.3
networkx==3.3
nomic==3.0.25
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.1
packaging==23.2
pandas==2.2.2
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.1
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
posthog==3.5.0
protobuf==4.25.5
ptyprocess==0.7.0
pyarrow==16.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.21.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
sse-starlette==1.8.2
starlette==0.37.2
streamlit==1.40.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.4
torch==2.2.2
tornado==6.4.1
tqdm==4.66.2
transformers==4.40.1
trove-classifiers==2024.4.10
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.23.2
uvloop==0.19.0
virtualenv==20.26.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
xattr==1.1.0
yarl==1.9.4
zstandard==0.22.0
